pyproject.toml
[tool.poetry]
name = "django-social-network"
version = "0.1.0"
description = ""
authors = ["Ali Shokrollahi <ali.shokrollahi.pro@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
django-mail-templated = "^2.6.5"
coverage = "^7.6.1"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
pytest-django = "^4.8.0"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
